refactor(macos): extract queue inspector & actions from AppModel into AppModel+Queue.swift
Continues the AppModel god-object split. Relocates the Up-Next inspector
projection and the queue-editing actions (move/remove/clear, remove-with-undo,
jump-to-index, reorder + core-queue sync) — the last method sections in the
class body — into a new `extension AppModel` file.

Pure relocation: the 263 moved lines are byte-identical to what was removed.
AppModel.swift drops 5,227 -> 4,964 lines; the class body now ends right after
its stored state.

One access promotion: sessionPlayHistoryLimit (private let -> internal),
referenced by the moved inspector projection. syncCoreQueueAfterReorder is
self-contained and stays private in the new file; the queue stored state was
already internal.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
